+++ custom_components/dmx_monitor/const.py
@@ -1,5 +1,5 @@
 DOMAIN = "dmx_monitor"
-VERSION = "0.13.0"
+VERSION = "0.13.1"
 
 CONF_INTERFACE = "interface"
 CONF_UNIVERSES = "universes"